separate swoole_event.
