+++ package.xml
@@ -157,6 +157,7 @@
 			<file role="src" name="swoole_server.c" />
 			<file role="src" name="swoole_lock.c" />
 			<file role="src" name="swoole_client.c" />
+            <file role="src" name="swoole_event.c" />
 			<file role="src" name="swoole_async.c" />
 			<file role="src" name="swoole_process.c" />
 			<file role="src" name="swoole_table.c" />
