+++ swoole_client.c
@@ -34,12 +34,6 @@
 #define php_sw_client_onClose       "onClose"
 #define php_sw_client_onError       "onError"
 
-typedef struct
-{
-	zval *cb_read;
-	zval *cb_write;
-	zval *socket;
-} swoole_reactor_fd;
 
 static char *php_sw_callbacks[PHP_CLIENT_CALLBACK_NUM] =
 {
@@ -71,10 +65,6 @@ static int php_swoole_client_event_add(zval *sock_array, fd_set *fds, int *max_f
 static int php_swoole_client_event_loop(zval *sock_array, fd_set *fds TSRMLS_DC);
 static int php_swoole_client_close(zval **zobject, int fd TSRMLS_DC);
 
-static int php_swoole_event_onRead(swReactor *reactor, swEvent *event);
-static int php_swoole_event_onWrite(swReactor *reactor, swEvent *event);
-static int php_swoole_event_onError(swReactor *reactor, swEvent *event);
-
 static int php_swoole_client_onRead(swReactor *reactor, swEvent *event);
 static int php_swoole_client_onWrite(swReactor *reactor, swEvent *event);
 static int php_swoole_client_onError(swReactor *reactor, swEvent *event);
@@ -84,7 +74,6 @@ static int swoole_client_error_callback(zval *zobject, swEvent *event, int error
 static void php_swoole_onTimeout(swTimer *timer, void *data);
 static void php_swoole_onTimerInterval(swTimer *timer, int interval);
 
-static int swoole_convert_to_fd(zval **fd);
 static swClient* swoole_client_create_socket(zval *object, char *host, int host_len, int port);
 
 void swoole_client_init(int module_number TSRMLS_DC)
@@ -307,25 +296,6 @@ static int php_swoole_client_onError(swReactor *reactor, swEvent *event)
 	return SW_OK;
 }
 
-static int php_swoole_event_onError(swReactor *reactor, swEvent *event)
-{
-    TSRMLS_FETCH_FROM_CTX(sw_thread_ctx ? sw_thread_ctx : NULL);
-
-    int error;
-    socklen_t len = sizeof(error);
-
-    if (getsockopt(event->fd, SOL_SOCKET, SO_ERROR, &error, &len) < 0)
-    {
-        php_error_docref(NULL TSRMLS_CC, E_WARNING, "swoole_event->onError[1]: getsockopt[sock=%d] failed. Error: %s[%d]", event->fd, strerror(errno), errno);
-    }
-    if (error != 0)
-    {
-        php_error_docref(NULL TSRMLS_CC, E_WARNING, "swoole_event->onError[1]: socket error. Error: %s [%d]", strerror(error), error);
-    }
-    SwooleG.main_reactor->del(SwooleG.main_reactor, event->fd);
-    return SW_OK;
-}
-
 static int php_swoole_client_onWrite(swReactor *reactor, swEvent *event)
 {
 	swClient *cli;
@@ -476,9 +446,7 @@ void php_swoole_check_reactor()
         SwooleG.main_reactor->setHandle(SwooleG.main_reactor, (SW_FD_USER + 1) | SW_EVENT_WRITE, php_swoole_client_onWrite);
         SwooleG.main_reactor->setHandle(SwooleG.main_reactor, (SW_FD_USER + 1) | SW_EVENT_ERROR, php_swoole_client_onError);
 
-		SwooleG.main_reactor->setHandle(SwooleG.main_reactor, SW_FD_USER | SW_EVENT_READ , php_swoole_event_onRead);
-		SwooleG.main_reactor->setHandle(SwooleG.main_reactor, SW_FD_USER | SW_EVENT_WRITE, php_swoole_event_onWrite);
-		SwooleG.main_reactor->setHandle(SwooleG.main_reactor, SW_FD_USER | SW_EVENT_ERROR, php_swoole_event_onError);
+        swoole_event_init();
 
 		SwooleWG.reactor_init = 1;
 	}
@@ -618,68 +586,6 @@ static void php_swoole_onTimerInterval(swTimer *timer, int interval)
 	zval_ptr_dtor(&zinterval);
 }
 
-static int php_swoole_event_onRead(swReactor *reactor, swEvent *event)
-{
-	zval *retval;
-	zval **args[1];
-	swoole_reactor_fd *fd;
-
-	TSRMLS_FETCH_FROM_CTX(sw_thread_ctx ? sw_thread_ctx : NULL);
-
-	if (zend_hash_find(&php_sw_event_callback, (char *)&(event->fd), sizeof(event->fd), (void**)&fd) != SUCCESS)
-	{
-		php_error_docref(NULL TSRMLS_CC, E_WARNING, "swoole_event: onRead not found");
-		return SW_ERR;
-	}
-
-	args[0] = &fd->socket;
-
-	if (call_user_function_ex(EG(function_table), NULL, fd->cb_read, &retval, 1, args, 0, NULL TSRMLS_CC) == FAILURE)
-	{
-		php_error_docref(NULL TSRMLS_CC, E_WARNING, "swoole_event: onRead handler error");
-		return SW_ERR;
-	}
-	if (retval != NULL)
-	{
-		zval_ptr_dtor(&retval);
-	}
-	return SW_OK;
-}
-
-static int php_swoole_event_onWrite(swReactor *reactor, swEvent *event)
-{
-    zval *retval;
-    zval **args[1];
-    swoole_reactor_fd *fd;
-
-    TSRMLS_FETCH_FROM_CTX(sw_thread_ctx ? sw_thread_ctx : NULL);
-
-    if (zend_hash_find(&php_sw_event_callback, (char *) &(event->fd), sizeof(event->fd), (void**) &fd) != SUCCESS)
-    {
-        php_error_docref(NULL TSRMLS_CC, E_WARNING, "swoole_event: onWrite not found");
-        return SW_ERR;
-    }
-
-    if (!fd->cb_write)
-    {
-        return swReactor_onWrite(reactor, event);
-    }
-
-    args[0] = &fd->socket;
-
-    if (call_user_function_ex(EG(function_table), NULL, fd->cb_write, &retval, 1, args, 0, NULL TSRMLS_CC) == FAILURE)
-    {
-        php_error_docref(NULL TSRMLS_CC, E_WARNING, "swoole_event: onWrite handler error");
-        return SW_ERR;
-    }
-
-    if (retval != NULL)
-    {
-        zval_ptr_dtor(&retval);
-    }
-    return SW_OK;
-}
-
 static swClient* swoole_client_create_socket(zval *object, char *host, int host_len, int port)
 {
 	zval *ztype, *zres;
@@ -774,60 +680,6 @@ static swClient* swoole_client_create_socket(zval *object, char *host, int host_
 	return cli;
 }
 
-static int swoole_convert_to_fd(zval **fd)
-{
-	php_stream *stream;
-	int socket_fd;
-
-	TSRMLS_FETCH_FROM_CTX(sw_thread_ctx ? sw_thread_ctx : NULL);
-
-#ifdef SWOOLE_SOCKETS_SUPPORT
-	php_socket *php_sock;
-#endif
-	if (Z_TYPE_PP(fd) == IS_RESOURCE)
-	{
-		if (ZEND_FETCH_RESOURCE_NO_RETURN(stream, php_stream *, fd, -1, NULL, php_file_le_stream()))
-		{
-			if (php_stream_cast(stream, PHP_STREAM_AS_FD_FOR_SELECT | PHP_STREAM_CAST_INTERNAL, (void* )&socket_fd, 1)
-					!= SUCCESS || socket_fd < 0)
-			{
-				return SW_ERR;
-			}
-		}
-		else
-		{
-#ifdef SWOOLE_SOCKETS_SUPPORT
-			if (ZEND_FETCH_RESOURCE_NO_RETURN(php_sock, php_socket *, fd, -1, NULL, php_sockets_le_socket()))
-			{
-				socket_fd = php_sock->bsd_socket;
-
-			}
-			else
-			{
-				php_error_docref(NULL TSRMLS_CC, E_WARNING, "fd argument must be either valid PHP stream or valid PHP socket resource");
-				return SW_ERR;
-			}
-#else
-			php_error_docref(NULL TSRMLS_CC, E_WARNING, "fd argument must be valid PHP stream resource");
-			return SW_ERR;
-#endif
-		}
-	}
-	else if (Z_TYPE_PP(fd) == IS_LONG)
-	{
-		socket_fd = Z_LVAL_PP(fd);
-		if (socket_fd < 0)
-		{
-			php_error_docref(NULL TSRMLS_CC, E_WARNING, "invalid file descriptor passed");
-			return SW_ERR;
-		}
-	}
-	else
-	{
-		return SW_ERR;
-	}
-	return socket_fd;
-}
 
 PHP_FUNCTION(swoole_timer_add)
 {
@@ -890,257 +742,6 @@ PHP_FUNCTION(swoole_timer_del)
     RETURN_TRUE;
 }
 
-PHP_FUNCTION(swoole_event_add)
-{
-	zval *cb_read = NULL;
-	zval *cb_write = NULL;
-	zval **fd;
-	char *func_name = NULL;
-	long event_flag = 0;
-
-	if (zend_parse_parameters(ZEND_NUM_ARGS() TSRMLS_CC, "Z|zzl", &fd, &cb_read, &cb_write, &event_flag) == FAILURE)
-	{
-		return;
-	}
-
-#ifdef ZTS
-	if (sw_thread_ctx == NULL)
-	{
-		TSRMLS_SET_CTX(sw_thread_ctx);
-	}
-#endif
-
-	if ((cb_read == NULL && cb_write == NULL) || (ZVAL_IS_NULL(cb_read) && ZVAL_IS_NULL(cb_write)))
-	{
-		php_error_docref(NULL TSRMLS_CC, E_WARNING, "no read or write event callback.");
-		RETURN_FALSE;
-	}
-
-    int socket_fd = swoole_convert_to_fd(fd);
-    if (socket_fd < 0)
-    {
-        php_error_docref(NULL TSRMLS_CC, E_WARNING, "unknow type.");
-        RETURN_FALSE;
-    }
-
-	swoole_reactor_fd event;
-	event.socket = *fd;
-	event.cb_read = cb_read;
-	event.cb_write = cb_write;
-
-	zval_add_ref(&event.socket);
-
-	if (cb_read!= NULL && !ZVAL_IS_NULL(cb_read))
-	{
-		if (!zend_is_callable(cb_read, 0, &func_name TSRMLS_CC))
-		{
-			php_error_docref(NULL TSRMLS_CC, E_ERROR, "Function '%s' is not callable", func_name);
-			efree(func_name);
-			RETURN_FALSE;
-		}
-		zval_add_ref(&event.cb_read);
-	}
-
-	if (cb_write!= NULL && !ZVAL_IS_NULL(cb_write))
-	{
-		if (!zend_is_callable(cb_write, 0, &func_name TSRMLS_CC))
-		{
-			php_error_docref(NULL TSRMLS_CC, E_ERROR, "Function '%s' is not callable", func_name);
-			efree(func_name);
-			RETURN_FALSE;
-		}
-		zval_add_ref(&event.cb_write);
-	}
-
-	if (zend_hash_update(&php_sw_event_callback, (char *)&socket_fd, sizeof(socket_fd), &event, sizeof(swoole_reactor_fd), NULL) == FAILURE)
-	{
-		php_error_docref(NULL TSRMLS_CC, E_WARNING, "swoole_event_add add to hashtable failed");
-		RETURN_FALSE;
-	}
-
-	php_swoole_check_reactor();
-	swSetNonBlock(socket_fd); //must be nonblock
-
-	if (SwooleG.main_reactor->add(SwooleG.main_reactor, socket_fd, SW_FD_USER | event_flag) < 0)
-	{
-		php_error_docref(NULL TSRMLS_CC, E_WARNING, "swoole_event_add failed.");
-		RETURN_FALSE;
-	}
-
-	php_swoole_try_run_reactor();
-	RETURN_LONG(socket_fd);
-}
-
-PHP_FUNCTION(swoole_event_write)
-{
-    zval **fd;
-    char *data;
-    int len;
-
-    if (zend_parse_parameters(ZEND_NUM_ARGS() TSRMLS_CC, "Zs", &fd, &data, &len) == FAILURE)
-    {
-        return;
-    }
-
-    if (len <= 0)
-    {
-        php_error_docref(NULL TSRMLS_CC, E_WARNING, "data empty.");
-        RETURN_FALSE;
-    }
-
-    if (!SwooleG.main_reactor)
-    {
-        php_error_docref(NULL TSRMLS_CC, E_WARNING, "reactor no ready, cannot write.");
-        RETURN_FALSE;
-    }
-
-    int socket_fd = swoole_convert_to_fd(fd);
-    if (socket_fd < 0)
-    {
-        php_error_docref(NULL TSRMLS_CC, E_WARNING, "unknow type.");
-        RETURN_FALSE;
-    }
-
-    if (SwooleG.main_reactor->write(SwooleG.main_reactor, socket_fd, data, len) < 0)
-    {
-        RETURN_FALSE;
-    }
-    else
-    {
-        RETURN_TRUE;
-    }
-}
-
-PHP_FUNCTION(swoole_event_set)
-{
-	zval *cb_read = NULL;
-	zval *cb_write = NULL;
-	zval **fd;
-	swoole_reactor_fd *ev_set;
-	char *func_name = NULL;
-	long event_flag = 0;
-
-	if (zend_parse_parameters(ZEND_NUM_ARGS() TSRMLS_CC, "Z|zzl", &fd, &cb_read, &cb_write, &event_flag) == FAILURE)
-	{
-		return;
-	}
-
-#ifdef ZTS
-	if (sw_thread_ctx == NULL)
-	{
-		TSRMLS_SET_CTX(sw_thread_ctx);
-	}
-#endif
-
-    int socket_fd = swoole_convert_to_fd(fd);
-    if (socket_fd < 0)
-    {
-        php_error_docref(NULL TSRMLS_CC, E_WARNING, "unknow type.");
-        RETURN_FALSE;
-    }
-
-	if (zend_hash_find(&php_sw_event_callback, (char *)&socket_fd, sizeof(socket_fd), (void **)&ev_set) != SUCCESS)
-	{
-		php_error_docref(NULL TSRMLS_CC, E_WARNING, "swoole_event: no such socket [fd=%d].", socket_fd);
-		RETURN_FALSE;
-	}
-
-    if (cb_read != NULL && !ZVAL_IS_NULL(cb_read))
-    {
-        if (!zend_is_callable(cb_read, 0, &func_name TSRMLS_CC))
-        {
-            php_error_docref(NULL TSRMLS_CC, E_ERROR, "Function '%s' is not callable", func_name);
-            efree(func_name);
-            RETURN_FALSE;
-        }
-        else
-        {
-            ev_set->cb_read = cb_read;
-            zval_add_ref(&cb_read);
-        }
-    }
-
-    if (cb_write != NULL && !ZVAL_IS_NULL(cb_write))
-    {
-        if (!zend_is_callable(cb_write, 0, &func_name TSRMLS_CC))
-        {
-            php_error_docref(NULL TSRMLS_CC, E_ERROR, "Function '%s' is not callable", func_name);
-            efree(func_name);
-            RETURN_FALSE;
-        }
-        else
-        {
-            ev_set->cb_write = cb_write;
-            zval_add_ref(&cb_write);
-        }
-    }
-
-	if ((event_flag & SW_EVENT_READ) && ev_set->cb_read == NULL)
-	{
-		php_error_docref(NULL TSRMLS_CC, E_WARNING, "swoole_event: no read callback.");
-	 	RETURN_FALSE;
-	}
-
-	if ((event_flag & SW_EVENT_WRITE) && ev_set->cb_write == NULL)
-	{
-		php_error_docref(NULL TSRMLS_CC, E_WARNING, "swoole_event: no write callback.");
-		RETURN_FALSE;
-	}
-
-	if (SwooleG.main_reactor->set(SwooleG.main_reactor, socket_fd, SW_FD_USER | event_flag) < 0)
-	{
-		php_error_docref(NULL TSRMLS_CC, E_WARNING, "swoole_event_set failed.");
-		RETURN_FALSE;
-	}
-
-	RETURN_TRUE;
-}
-
-PHP_FUNCTION(swoole_event_del)
-{
-	zval **fd;
-	if (zend_parse_parameters(ZEND_NUM_ARGS() TSRMLS_CC, "Z", &fd) == FAILURE)
-	{
-		return;
-	}
-	int socket_fd = swoole_convert_to_fd(fd);
-	if (socket_fd < 0)
-	{
-		php_error_docref(NULL TSRMLS_CC, E_WARNING, "unknow type.");
-		RETURN_FALSE;
-	}
-	SW_CHECK_RETURN(SwooleG.main_reactor->del(SwooleG.main_reactor, socket_fd));
-}
-
-PHP_FUNCTION(swoole_event_exit)
-{
-	if (SwooleWG.in_client == 1)
-	{
-		//stop reactor
-		SwooleG.running = 0;
-	}
-}
-
-PHP_FUNCTION(swoole_event_wait)
-{
-    if (SwooleWG.in_client == 1 && SwooleWG.reactor_ready == 0 && SwooleG.running)
-    {
-        SwooleWG.reactor_ready = 1;
-
-#ifdef HAVE_SIGNALFD
-        if (SwooleG.main_reactor->check_signalfd)
-        {
-            swSignalfd_setup(SwooleG.main_reactor);
-        }
-#endif
-
-        int ret = SwooleG.main_reactor->wait(SwooleG.main_reactor, NULL);
-        if (ret < 0)
-        {
-            php_error_docref(NULL TSRMLS_CC, E_ERROR, "reactor wait failed. Error: %s [%d]", strerror(errno), errno);
-        }
-    }
-}
 
 PHP_METHOD(swoole_client, __construct)
 {
